job_notifier/requirements.txt
aiohttp==3.9.1
beautifulsoup4==4.12.2
python-telegram-bot==20.7
schedule==1.2.1
requests==2.31.0
duckduckgo-search==4.1.1
